Mantenedor insumo casi listo, falta validar y arreglar aspectos de diseño

+++ src/main/java/com/restaurante/bacon/dto/Insumo.java
@@ -39,12 +39,12 @@ public class Insumo implements Serializable {
     @Id
     @Basic(optional = false)
     //declarar que el id se usa con una secuencia 
-    @GeneratedValue(strategy=GenerationType.SEQUENCE, generator="SEC_INS")
-    @SequenceGenerator(name="SEC_INS",sequenceName="SEC_INSUMO", allocationSize=1 )
+    @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "SEC_INS")
+    @SequenceGenerator(name = "SEC_INS", sequenceName = "SEC_INSUMO", allocationSize = 1)
     //declarar la secuencia 
     @NotNull
     @Column(name = "ID_INSUMO")
-    private BigDecimal idInsumo;
+    private Integer idInsumo;
     @Basic(optional = false)
     @NotNull
     @Size(min = 1, max = 60)
@@ -77,16 +77,15 @@ public class Insumo implements Serializable {
     @Size(min = 1, max = 255)
     @Column(name = "FOTO_INSUMO")
     private String fotoInsumo;
-   
 
     public Insumo() {
     }
 
-    public Insumo(BigDecimal idInsumo) {
+    public Insumo(Integer idInsumo) {
         this.idInsumo = idInsumo;
     }
 
-    public Insumo(BigDecimal idInsumo, String nombreInsumo, String descripcionInsumo, BigInteger stockInsumo, String unidadMedidaInsumo, BigInteger minimoStockInsumo, BigInteger maximoStockInsumo, String fotoInsumo) {
+    public Insumo(Integer idInsumo, String nombreInsumo, String descripcionInsumo, BigInteger stockInsumo, String unidadMedidaInsumo, BigInteger minimoStockInsumo, BigInteger maximoStockInsumo, String fotoInsumo) {
         this.idInsumo = idInsumo;
         this.nombreInsumo = nombreInsumo;
         this.descripcionInsumo = descripcionInsumo;
@@ -97,11 +96,11 @@ public class Insumo implements Serializable {
         this.fotoInsumo = fotoInsumo;
     }
 
-    public BigDecimal getIdInsumo() {
+    public Integer getIdInsumo() {
         return idInsumo;
     }
 
-    public void setIdInsumo(BigDecimal idInsumo) {
+    public void setIdInsumo(Integer idInsumo) {
         this.idInsumo = idInsumo;
     }
 
@@ -161,8 +160,6 @@ public class Insumo implements Serializable {
         this.fotoInsumo = fotoInsumo;
     }
 
-    
-
     @Override
     public int hashCode() {
         int hash = 0;
@@ -183,5 +180,4 @@ public class Insumo implements Serializable {
         return true;
     }
 
-    
 }
